auth refractoring
