+++ controllers/users.js
@@ -3,9 +3,9 @@
 
 import { Router } from "express";
 import User from "../models/user.js";
-import newDate from "../functions/date.js";
+import dateLib from "../libraries/date_lib.js";
 import bcrypt from "bcrypt";
-import validators from "../functions/validators.js";
+import validators from "../libraries/validators.js";
 
 const usersRouter = Router();
 
@@ -36,7 +36,7 @@ usersRouter.post("/", async (req, res) => {
 
   const saltRounds = 10;
   const passwordHash = await bcrypt.hash(password, saltRounds);
-  const created = newDate();
+  const created = dateLib.newDateEST();
 
   const user = new User({
     username,
